Add switch for simplified log format
We made a change previously to simplify the log format and coloring
for console/stdout output. (I believe this was so that subcommands
like 'build' and 'deploy' produced more user friendly output.) This
is a bit too simple though since it is still useful to have the full
log format when running master and slave services.

We can have it both ways! This adds a flag to the configure_logging
call that will switch between the standard and simplfied log format.