Disable argparse argument abbreviations
The argparse library does prefix matching by default. For example, a
user can specify "--master" at the command line instead of
"--master-url", as long as the prefix is unambiguous.

Prefix matching is actually an undesired default behavior for us as it
creates the potential for unintended breaking changes just by adding a
new command-line argument.

For example, if a user uses the argument "--master" to specify a value
for "--master-url", and later we add a new argument named
"--master-port", that change will break the user script that used
"--master".

This change disables prefix matching in argparse.