Implement max_executors_per_slave job conf setting
We currently allow users to specify a "max_executors" setting on their
jobs. This is useful to prevent one job from "overparallelizing" and
monopolizing an entire cluster.

This change implements an additional job config setting called
"max_executors_per_slave". This is useful, for example, on jobs which
*can* be parallelized across multiple machines, but *cannot* be
parallelized with multiple processes on a single machine. (That problem
can now be solved by setting max_executors_per_slave to 1.)