Update slave allocation increment value
Number of executors for a build was being incremented by
the number of executors on a slave during allocation.

If the max executors per slave, is set to a value
significantly less than the number of executors on the slave,
a build could end up with far fewer allocated slaves then expected.

This commit makes it so that the number of executors on the build
is incrmented by max executors per slave if it is less than the
number of executors available on the slave.