Prevent clusterrunner zombie processes
This change adds a method to the main thread that will start a timer
of 10 seconds right before the main thread finishes. After 10 seconds,
the application will dump some debug info to the logs and then send
itself a SIGKILL signal.

The dumped debug info includes the stack traces of all currently
running threads.

This is to address an issue where we see the application shutdown the
web server but the process continues to run. This does not affect
deployment of newer versions but just causes us to have some old
zombie clusterrunner processes that we have to kill manually. This
change should prevent those zombies.