Increase robustness of subprocess communication
Issue #57 describes an exception we've been seeing bubble out of
`Popen.communicate()`. This happens when we're executing a command for a
subjob, and the exception ends up shutting down the slave.

This change attempts to prevent that specific exception, and also adds
some error handling so that similar issues will only fail the current
build and not bring down the entire slave.

Specifically, this changes the way we're retrieving stdout and stderr
from subprocesses. Previously we used pipes to the subprocess, but this
changes output to be written to and read from a temp file. This allows
us to use Popen's `wait()` instead of `communicate()` and will hopefully
prevent the above exception.

This should fix #57.