Builds should never hang if all subjobs complete
Currently, if a slave fails during setup_build or teardown_build, the
build will be stuck in BUILDING state forever, even if all subjobs
complete successfully on other slaves. (Issue #65)

The reason is that we have code that prevents a build from finishing
until all slaves ever assigned to that build have successfully
completed teardown_build commands. This is actually unnecessary - a
build should be considered complete as soon as its last subjob
completes.

This change removes the code that blocks a build from finishing until
teardown_build is complete. It also clears a slave's `current_build_id`
field when a slave disconnects (since a disconnected slave is not
working on any build.)

Fixes #65.