Fix issue with too many open files on master
This change fixes #69.

There are three pieces to this change:
- Separate out the logic that uses pexpect and make it execute in a
  Python subprocess. Pexpect's `child.expect()` raises an exception when
  the process it is launched from has too many open files. (See #69 for
  more details on the issue.)
- Change the version of pexpect in our requirements file to pull a
  a specific commit directly from the pexpect repo. This specific commit
  includes a bugfix for running pexpect in a Python subprocess.
- Add a functional test that tests the git project type so that the new
  multiprocess logic is covered. (Our existing functional tests only
  test the directory type.) This test clones the ClusterRunner demo
  project. (This is skipped for now since it fails on Travis-CI due to
  unrelated ssh issues.)

Because of the way the alternate pexpect version is pulled in, we'll
need to run `pip install --upgrade -r requirements.txt` to make sure
that the official 3.3 release is replaced with the snapshot.

I've also temporarily disabled the kill_event in git.py since we want
to get this change out quickly. I'll follow up with another commit to
reenable this.