Make multiprocessing usages compatible with freeze
This commit makes corrections to the changes merged in #70 and
reverted in #73. Those changes were incompatible with freezing since
they used the 'spawn' method of starting a subprocess. After a good
amount of investigation, it looks like the 'spawn' multiprocessing
start method is incompatible with freezing specifically on unix.

This commit switches the start method back to the default (fork) and
adds necessary changes to get this working. Since forked processes
inherit file descriptors of their parent process, we iterate through
the open file descriptors in the child process and close the ones we
don't need (TCP connections).