More detailed versioning scheme
This change takes our current versioning scheme and adds more
info to the version string for branches and uncommitted changes. The
intention here is to have the version string give us a clear
distinction between trunk, feature branch, and development builds.

The version scheme stays the same for trunk commits (e.g., `0.5.155`).
Trunk commits are defined as being in the first-parent git log of
`origin/master`, which can be viewed with:
   `git log --first-parent --oneline origin/master`

For non-trunk commits, the git short hash is appended to the version
number, for example: `0.5.156-cac0414`.

When there are modified files in the repo that have not yet been
committed, `-mod` is appended to the end of the version string. For
example: `0.5.156-cac0414-mod`.

This scheme maintains the expected version ordering using setuptools'
versioning guidelines. More info on this can be found at
https://pythonhosted.org/setuptools/setuptools.html#specifying-your-project-s-version