Add clusterrunner.yaml to repo
This enables running the ClusterRunner unit tests via ClusterRunner! On
my laptop running a slave with 8 executors cuts down the amount of time
it takes to run the unit tests by about half.

Other changes in this commit to support this:
- Add __init__.py to a couple test directories where the file was
missing (meaning those tests weren't even being run)
- Fix or remove tests in app/util/shell that have been silently broken
because up until now they weren't being run
- Add a test that will ensure an __init__.py file is added to all test
directories in the future
- Add a command to .travis.yml to run unit tests via ClusterRunner.
This works as both a way to run our unit tests and as a real-world
functional test.