Add an executor_index for the entire build
There is a use case for functional tests where a set
of logins are generated once, then each test executor
in a build needs to use a login that is not used by
any other executor in the build.

This adds a BUILD_EXEUCTOR_INDEX environment param
which stores a sequential executor id, unique across
the whole build. This can be used to index into a list
of logins created by the user.

Also adds a MACHINE_EXECUTOR_INDEX which is the same as
EXECUTOR_INDEX.