Fix git fetching
There is a race condition right now where if a git branch is specified
in a build request, slaves will pull from the master using that branch
name (whose head may change when the master handles future builds).

This change creates a local ref on the master whenever a git build is
requested. This ref is then passed to the slaves instead of the user-
specified branch name so that the slaves are guaranteed to always fetch
the same commit.

As part of this change, I also refactored how the slave build parameter
overriding works.