Translate 'localhost' to hostname when connecting slaves
When deploying clusterrunner with master_host set to `localhost`, all
the slaves will actually try to connect to `localhost:43000`, which, of
course, is going to fail. This change translates 'localhost' to the
actual host name of the master before it constructs the command and send
it to slave to connect to master.

This fixes #144