Instead of writing stdout/ stderr to temp files, write to artiact path
On the slave service, we are writing stdout and stderr to two separate
temporary files, and later copying the contents of them to its final
path in the artifact directory.

This will not only be slightly more efficient, but it allows us to
implement returning the console output of in-progress builds from
slaves.