Refactor and clean up tests for Build object
The Build class is the largest single class and the most complex object
in the application. Future planned refactorings will help this, but in
the meantime it is hard and usually verbose to create realistic build
objects for testing.

This change refactors some of the Build tests to use a build creation
helper that creates more realistic build objects. It has a side benefit
of making these tests more readable and maintainable.

The tradeoff is that these tests are now more "integration-y", but it
makes sense to move in this direction considering the complexity of the
Build class itself.