Update timing data keys, not overwrite it entirely.
This will be necessary when we implement overridden atoms from build
requests. In such a build request, we will not have ran all the atoms,
and if we were to keep the existing behavior, we would wipe out all of
the atom timing data except for the select few that were manually
specified in the build request.

This method also leaves the door open in the future if we want to always
update the timing data, regardless of overall build success/failure, but
only for successful atoms.