Enable Session-Id headers for slaves
The master already returns a session id in response headers, and the
API will raise a 412 if the Session-Id request header specifies a
differing session id.

This change moves that logic one level up so that slaves will also
start generating and returning a slave-specific session id in their
responses. The master now also tracks the session id for each slave.

A subsequent change will add verification when the master sends
requests to the slave to ensure that the master is talking to the
correct slave instance.