Verify slave session id on is_alive check
Previously when the master checked if a slave is alive, it only
checked if the slave service is responsive at that url. This change
makes the master also verify that the slave service that responds at
that url still has the same session id that the slave reported when
it connected to the master.

This will cause the `slave.is_alive()` check on the master to start
returning false if the slave responds but has the wrong session id.
Previously this would return true.

This is done by setting a header in the master's request to the
slave that includes the expected session id. If the slave sees this
and it does not match its current session id, it will return a 412
and the master will consider the slave with the old session id to be
offline.

This fixes #296. Now when the slave allocator does the `is_alive()`
call before allocating a slave to a build, it will correctly detect
when a slave's session id has changed and should be considered
offline.