Prevent immortal connection pools
The master keeps a connection pool for each connected slave. The
problem is that even when the slave disconnects/is taken offline
those pooled connections stay open. This isn't that bad unless
you're constantly connecting and disconnecting slaves.

At Box we constantly connect and disconnect slaves (recycling the
entire cluster once a night). Over several weeks the immortal
connection pools eventually cause the master process to hit a system
limit on the number of open file descriptors. :(

This change makes sure we clear the connection pool when a slave
disconnects.