Add Prometheus instrumentation
This change uses the Prometheus client SDK to add some instrumentation
to the master. So far, only HTTP request latency is instrumented.

There is also a new endpoint at /v1/metrics which exposes the metrics in
in Prometheus text exposition format. The metrics endpoint required a
new method in ClusterBaseHandler so that a raw text string could be
written.