Rename archive methods to reference tar
The current artifact archiving methods and variables only refer to
"artifact" and "compress". To address #329 we'd like to create a zip
archive file as well, so this change renames the current methods to
remove ambiguitiy between tar and (future) zip methods.

I also refactored `tar_directory` slightly so that it uses
`tar_directories`.