Deprecate build result tar.gz in favor of zip (#335)
This adds an artifacts.zip with exactly the same contents as the
existing artifacts.tar.gz. The tar.gz artifact format is now
deprecated and clients should update to use the zip file instead.

The benefit of using a zip is that it will be easier to fix a bug
around loading console output from completed builds. By having a zip
archive of the build results on the master, we'll be able to load
console output directly from the master without redirecting to the
slave that the build ran on. The fix for this issue is coming in a
follow-up change.

Fixes #329.