Add timeout to all outgoing http requests (#338)
Currently if a slave becomes unresponsive the master service will
freeze up for a long time (or forever). The slave allocation loop
will hang trying to contact a slave and no builds will be able to
start.

This adds a timeout to all requests between the master and slaves
and corresponding error handling. I made the timeout on
slave->master calls longer because we don't currently have great
metrics and have had issues with request load on the master in the
past (which we should improve separately).