Ignore zipfile artifact creation failures
We've seen an issue on our ClusterRunner production instance that
will cause zip creation to fail with a FileNotFoundError. Since
we're not yet depending on zip files, this just logs and ignores any
failures that occur.

This also moves zip creation to occur after tar.gz creation.
Hopefully we'll be able to use the successfully created tar.gz
artifact to debug issues with creating the zip artifact.