Pull logs from zip archive if it exists
Currently logs are only available on the slave once a build has
completed because the console output files are moved into an archive
file on the master. This means that if a slave is taken down that
logs for builds that ran on it will not be available via the console
output API.

This updates the logic to pull logs from the zip archive if it
exists on the master which will allow slaves to be taken offline
without losing log functionality.