Remove exception logging for ItemNotReadyError
Attempting to reach an item that isn't ready is basically a 404 and
there isn't really a need to log anything to the master in this case.

Also when attempting to reach an artifact that isn't ready, the
`clusterrunner build` command will continuously try to request it
until the client is explicitly stopped -- this will spam the master
logs.

Resolves: #342