Fail functional tests if builds are slow
We see some flakiness in CI related to builds not being in an
expected state at the end of a test. There are calls to block a test
until a build reaches a certain state, but if the build doesn't
reach that state within the timeout we currently just ignore that
and continue on. This change adds assertions so that tests will fail
immediately if a build is taking too long.

This is a step towards making sources of flakiness more obvious in
the functional tests.