Convert subjob start SafeThread to standard Thread (#384)
Sending a subjob to a slave is done asynchronously. Even though it's
bad if it fails, it shouldn't bring down the entire master service.

This is a quick fix because we've run into this issue multiple
times. I'll follow up with a change to clean this up and improve the
behavior in the error case.